Move test to pydantic
