+++ pydantic-core/tests/validators/test_url.py
@@ -1493,23 +1493,6 @@ def test_multi_url_build_hosts() -> None:
     )
 
 
-def test_multi_url_build_hosts_with_none_values() -> None:
-    """https://github.com/pydantic/pydantic/issues/13007"""
-
-    hosts: list[MultiHostHost] = [
-        {'host': 'host-1.com', 'password': 'pass', 'username': 'user', 'port': 27017},
-        {'host': 'host-2.com', 'password': None, 'username': None, 'port': 27017},
-    ]
-    url = MultiHostUrl.build(
-        scheme='mongodb',
-        hosts=hosts,
-        path='db',
-        query='replicaSet=xxx&authSource=admin',
-    )
-    assert str(url) == 'mongodb://user:pass@host-1.com:27017,host-2.com:27017/db?replicaSet=xxx&authSource=admin'
-    assert url.hosts() == hosts
-
-
 def test_multi_url_build_neither_host_and_hosts_set() -> None:
     with pytest.raises(ValueError):
         MultiHostUrl.build(
